// src/modules/shared/store/store-key.enum.ts
enum StoreKey {
  AppVersion = 'appVersion',
  BookmarkIdMappings = 'bookmarkIdMappings',
  Bookmarks = 'bookmarks',
  CheckForAppUpdates = 'checkForAppUpdates',
  DarkModeEnabled = 'darkModeEnabled',
  DisplayHelp = 'displayHelp',
  DisplayOtherSyncsWarning = 'displayOtherSyncsWarning',
  DisplayPermissions = 'displayPermissions',
  DisplaySearchBarBeneathResults = 'displaySearchBarBeneathResults',
  DisplayUpdated = 'displayUpdated',
  DefaultToFolderView = 'defaultToFolderView',
  InstallBackup = 'installBackup',
  LastUpdated = 'lastUpdated',
  Password = 'password',
  ServiceUrl = 'serviceUrl',
  SyncBookmarksToolbar = 'syncBookmarksToolbar',
  SyncEnabled = 'syncEnabled',
  SyncId = 'syncId',
  SyncVersion = 'syncVersion',
  TraceLog = 'traceLog'
}

export default StoreKey;